Fix services

+++ src/main/java/com/app/concessionario/services/ClienteService.java
@@ -7,7 +7,6 @@ import com.app.concessionario.mapper.ClienteMapper;
 import com.app.concessionario.repositories.ClienteRepository;
 import com.app.concessionario.repositories.ConcessionarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -25,7 +24,7 @@ public class ClienteService {
 
 //                                  CHIAMATE DTO
 
-//    chiamata get per tutti i clienti
+    //    chiamata get per tutti i clienti
     public List<ClienteDTO> getClientiDTO() {
         List<ClienteDTO> clientiDTO = new ArrayList<>();
         List<Cliente> clienti = clienteRepository.findAll();
@@ -36,59 +35,71 @@ public class ClienteService {
         return clientiDTO;
     }
 
-//    chiamata get per un singolo cliente
+    //    chiamata get per un singolo cliente
     public ClienteDTO getClienteDTO(Integer id) throws Exception {
-        List<Integer> clientiIds = clienteRepository.findAll().stream().map(Cliente::getId).toList();
-        if (!clientiIds.contains(id))
+        Optional<Cliente> c = clienteRepository.findById(id);
+        if (!c.isPresent())
             throw new Exception("L'id inserito non è corretto");
 
-        Optional<Cliente> c = clienteRepository.findById(id);
         return ClienteMapper.toDTO(c.get());
     }
 
-//    chiamata in post per creare un nuovo cliente
+    //    chiamata in post per creare un nuovo cliente
     public void addClienteDTO(ClienteDTO clienteDTO) throws Exception {
-        List<Integer> clientiIds = clienteRepository.findAll().stream().map(Cliente::getId).toList();
-        List<String> clientiMails = clienteRepository.findAll().stream().map(Cliente::getEmail).toList();
-        if (clientiIds.contains(clienteDTO.getId())) {
+        Optional<Cliente> c = clienteRepository.findById(clienteDTO.getId());
+        String mail = c.get().getEmail();
+
+        if (c.isPresent())
             throw new Exception("L'id del cliente inserito non è valido");
-        } else if (clientiMails.contains(clienteDTO.getEmail())) {
-            throw new Exception("La mail inserita è già presente");
-        } else if (clienteDTO.getEmail() == null) {
+
+        if (clienteDTO.getEmail() == null)
             throw new Exception("Il campo email non può essere vuoto");
-        } else if (clienteDTO.getNome().length() > 30) {
+
+        if (clienteDTO.getEmail().equalsIgnoreCase(mail))
+            throw new Exception("La mail inserita è già presente");
+
+        if (clienteDTO.getNome().length() > 30)
             throw new Exception("Il campo nome non può superare i 30 caratteri");
-        } else if (clienteDTO.getCognome().length() > 30) {
+
+        if (clienteDTO.getCognome().length() > 30)
             throw new Exception("Il campo cognome non può superare i 30 caratteri");
-        } else if (clienteDTO.getEmail().length() > 100) {
+
+        if (clienteDTO.getEmail().length() > 100)
             throw new Exception("Il campo email non può superare i 100 caratteri");
-        } else if (clienteDTO.getTelefono().length() > 100) {
+
+        if (clienteDTO.getTelefono().length() > 100)
             throw new Exception("Il campo telefono non può superare i 30 caratteri");
-        }
 
         List<Concessionario> concessionari = concessionarioRepository.findAll();
         Cliente newCliente = ClienteMapper.toEntity(clienteDTO, concessionari);
         clienteRepository.save(newCliente);
     }
 
-//    chiamata per update cliente
+    //    chiamata per update cliente
     public void updateClienteDTO(Integer id, ClienteDTO clienteDTO) throws Exception {
-        List<Integer> clientiIds = clienteRepository.findAll().stream().map(Cliente::getId).toList();
-        if (clientiIds.contains(id)) {
-            List<String> clientiMails = clienteRepository.findAll().stream().map(Cliente::getEmail).toList();
-            if (clientiMails.contains(clienteDTO.getEmail())) {
-                throw new Exception("La mail inserita è già presente");
-            } else if (clienteDTO.getEmail() == null) {
+        Optional<Cliente> c = clienteRepository.findById(id);
+
+        if (c.isPresent()) {
+            String mail = c.get().getEmail();
+
+            if (clienteDTO.getEmail() == null)
                 throw new Exception("Il campo email non può essere vuoto");
-            } else if (clienteDTO.getNome().length() > 30) {
+
+            if (clienteDTO.getEmail().equalsIgnoreCase(mail))
+                throw new Exception("La mail inserita è già presente");
+
+            if (clienteDTO.getNome().length() > 30)
                 throw new Exception("Il campo nome non può superare i 30 caratteri");
-            } else if (clienteDTO.getCognome().length() > 30) {
+
+            if (clienteDTO.getCognome().length() > 30)
                 throw new Exception("Il campo cognome non può superare i 30 caratteri");
-            } else if (clienteDTO.getEmail().length() > 100) {
+
+            if (clienteDTO.getEmail().length() > 100)
                 throw new Exception("Il campo email non può superare i 100 caratteri");
-            } else if (clienteDTO.getTelefono().length() > 100) {
+
+            if (clienteDTO.getTelefono().length() > 100)
                 throw new Exception("Il campo telefono non può superare i 30 caratteri");
-            }
+
             List<Concessionario> concessionari = concessionarioRepository.findAll();
             Cliente newCliente = ClienteMapper.toEntity(clienteDTO, concessionari);
             newCliente.setId(id);
@@ -99,12 +110,14 @@ public class ClienteService {
     }
 
 //    chiamata per eliminare un cliente
-    public void deleteCliente(Integer id) throws Exception {
-        List<Integer> clientiIds = clienteRepository.findAll().stream().map(Cliente::getId).toList();
-        if (!clientiIds.contains(id)) {
-            throw new Exception("Cliente selezionato non presente");
-        }
-    clienteRepository.deleteById(id);
-}
+        public void deleteCliente (Integer id) throws Exception {
+            Optional<Cliente> c = clienteRepository.findById(id);
 
-}
+            if (!c.isPresent()) {
+                throw new Exception("Cliente selezionato non presente");
+            }
+            clienteRepository.deleteById(id);
+        }
+
+    }
+
